CMake: Do not use GLOB for source files
Using a GLOB to collect lists of source files is not recommended,
instead explicitly list source files.

With a GLOB, when someone else adds a file or changes a file name,
this change will not be picked up by CMake unless it is explicitly
run again. This can cause weird linker errors due to missing symbols
and other hard to debug issues.