EbDefinitions: Do not disable all assertions
EbDefinitions has a weird redefinition of asserts that cause asserts
to not work at all when EbDefinitions was included after assert.h
or if only EbDefinitions was included but no assert.h, without
any warning.

This removes the code completely as it is completely unnecessary and
highly misleading to mess with C library provided macros.